Use flex method to layout a header
I separate a header into both left and right sides. I set the classes a-
mong ul, both sides, header to `display:flex` and center every items in
the container. Also, I keep the 8px space between everything and the ed-
ge of the header. The items are arranged horizontally.